requirements.txt
Django==1.9.6
djangorestframework==3.3.3
drf-nested-routers==0.11.1
django-rest-framework-social-oauth2==1.0.4
django-cors-headers==1.1.0
psycopg2cffi
